+++ examples/cube-no-framework/src/main.rs
@@ -99,14 +99,23 @@ fn main() {
     )
     .unwrap();
 
+    // Create the shader preprocessor with all the default shaders added.
+    let mut spp = rend3::ShaderPreProcessor::new();
+    rend3_routine::builtin_shaders(&mut spp);
+
     // Create the base rendergraph.
-    let base_rendergraph = rend3_routine::base::BaseRenderGraph::new(&renderer);
+    let base_rendergraph = rend3_routine::base::BaseRenderGraph::new(&renderer, &spp);
 
     let mut data_core = renderer.data_core.lock();
-    let pbr_routine = rend3_routine::pbr::PbrRoutine::new(&renderer, &mut data_core, &base_rendergraph.interfaces);
+    let pbr_routine =
+        rend3_routine::pbr::PbrRoutine::new(&renderer, &mut data_core, &spp, &base_rendergraph.interfaces);
     drop(data_core);
-    let tonemapping_routine =
-        rend3_routine::tonemapping::TonemappingRoutine::new(&renderer, &base_rendergraph.interfaces, preferred_format);
+    let tonemapping_routine = rend3_routine::tonemapping::TonemappingRoutine::new(
+        &renderer,
+        &spp,
+        &base_rendergraph.interfaces,
+        preferred_format,
+    );
 
     // Create mesh and calculate smooth normals based on vertices
     let mesh = create_mesh();