+++ rend3-routine/src/common/mod.rs
@@ -2,12 +2,10 @@
 
 mod interfaces;
 mod samplers;
-mod shaders;
 mod sorting;
 mod vertex;
 
 pub use interfaces::*;
 pub use samplers::*;
-pub use shaders::*;
 pub use sorting::*;
 pub use vertex::*;